Clean up refactoring of prefixes in to prefix file
